+++ docs/components/pages/PageApi.vue
@@ -1,0 +1,15 @@
+<template>
+<div>
+  <section-api></section-api>
+</div>
+</template>
+
+<script>
+import SectionApi from '../sections/SectionApi';
+
+export default {
+  components: {
+    SectionApi,
+  },
+};
+</script>