+++ docs/components/pages/PageHome.vue
@@ -1,0 +1,33 @@
+<template>
+  <div>
+    <section-intro></section-intro>
+    <section-attributes></section-attributes>
+    <section-highlights></section-highlights>
+    <section-indicators></section-indicators>
+    <section-content-styles></section-content-styles>
+    <section-date-picker></section-date-picker>
+    <section-date-picker-slot></section-date-picker-slot>
+  </div>
+</template>
+
+<script>
+import SectionIntro from '../sections/SectionIntro';
+import SectionAttributes from '../sections/SectionAttributes';
+import SectionContentStyles from '../sections/SectionContentStyles';
+import SectionHighlights from '../sections/SectionHighlights';
+import SectionIndicators from '../sections/SectionIndicators';
+import SectionDatePicker from '../sections/SectionDatePicker';
+import SectionDatePickerSlot from '../sections/SectionDatePickerSlot';
+
+export default {
+  components: {
+    SectionIntro,
+    SectionAttributes,
+    SectionContentStyles,
+    SectionHighlights,
+    SectionIndicators,
+    SectionDatePicker,
+    SectionDatePickerSlot,
+  },
+};
+</script>