+++ docs/components/pages/PageSetup.vue
@@ -1,0 +1,15 @@
+<template>
+<div>
+  <section-setup></section-setup>
+</div>
+</template>
+
+<script>
+import SectionSetup from '../sections/SectionSetup';
+
+export default {
+  components: {
+    SectionSetup,
+  },
+};
+</script>