+++ docs/components/sections/SectionHighlights.vue
@@ -19,7 +19,7 @@
           </b-tab-item>
           <!--Highlights Example Code-->
           <b-tab-item label='Example Code'>
-            <pre v-highlight><code class='html'>{{ exHighlightsCode }}</code></pre>
+            <code-block :code='exHighlightsCode'></code-block>
           </b-tab-item>
         </b-tabs>
       </div>