+++ docs/components/sections/SectionIntro.vue
@@ -9,7 +9,7 @@
             <b-tab-item label='Overview'>
               <div class='content'>
                 <p>
-                  V-Calendar is a lightweight and dependency-free Vue.js plugin for displaying simple annotated calendars.
+                  V-Calendar is a clean and lightweight Vue.js plugin for displaying simple, annotated calendars.
                 </p>
                 <p>
                   It was designed from the outset with extendability and customization in mind.
@@ -22,13 +22,48 @@
             </b-tab-item>
             <!--Intro Example Code-->
             <b-tab-item label='Example Code'>
-              <pre v-highlight><code class='html'>{{ exIntroCode }}</code></pre>
+              <code-block :code='exIntroCode'></code-block>
+            </b-tab-item>
+            <!--Intro Options-->
+            <b-tab-item label='Options' icon='gear'>
+              <b-field>
+                <b-switch v-model='isDoublePaned'>Double Paned</b-switch>
+              </b-field>
+              <b-field label='Title Position'>
+                <p class='control'>
+                  <b-radio v-model='titlePosition' native-value='left'>Left</b-radio>
+                  <b-radio v-model='titlePosition' native-value=''>Center</b-radio>
+                  <b-radio v-model='titlePosition' native-value='right'>Right</b-radio>
+                </p>
+              </b-field>
+              <b-field label='Title Transition'>
+                <p class='control'>
+                  <b-radio v-model='titleTransition' native-value='none'>None</b-radio>
+                  <b-radio v-model='titleTransition' native-value='slide'>Slide</b-radio>
+                  <b-radio v-model='titleTransition' native-value='fade'>Fade</b-radio>
+                </p>
+              </b-field>
+              <b-field label='Weeks Transition'>
+                <p class='control'>
+                  <b-radio v-model='weeksTransition' native-value='none'>None</b-radio>
+                  <b-radio v-model='weeksTransition' native-value='slide'>Slide</b-radio>
+                  <b-radio v-model='weeksTransition' native-value='fade'>Fade</b-radio>
+                </p>
+              </b-field>
+              <br />
+              <b-field>
+                <p class='is-size-7'><strong>Note:</strong> Not all options available are displayed</p>
+              </b-field>
             </b-tab-item>
           </b-tabs>
         </div>
         <div class='column'>
           <div class='example-container'>
-            <ex-intro>
+            <ex-intro
+              :is-double-paned='isDoublePaned'
+              :title-position='titlePosition'
+              :title-transition='titleTransition'
+              :weeks-transition='weeksTransition'>
             </ex-intro>
           </div>
         </div>
@@ -50,10 +85,10 @@ export default {
   data() {
     return {
       exIntroCode: ExIntroCode,
-      isDoublePaned: false,
+      isDoublePaned: true,
+      titlePosition: '',
       titleTransition: 'slide',
       weeksTransition: 'slide',
-      transitions: ['slide', 'fade', 'none'],
     };
   },
 };