+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add tap to dismiss popover
 * Add color blending for days not in month (transparent layers are visually buggy and not as performant)
 * Add tint color (with blending) for easy date picker styling
+* Add support for custom title position (left, center, right)
 
 ## 0.0.4